refactor(a2a): address review — extract shared suggestTools helper, cap method echo, declare text/plain output
Greptile findings on #4838 + #4824, all taken:

- api/_agent-tool-suggest.ts (route-less underscore helper, like
  api/_crypto.js): suggestTools + tokenizer move out of api/a2a.ts so the
  /ask cold-start no longer executes a2a's module-level rate-limit guard —
  a future '/api/a2a' policy removal can't break /ask with a misleading
  '[a2a]' error. a2a.ts re-exports for a stable import surface.
- a2a -32601 branch: cap the echoed method name at 100 chars (reflected
  response amplification).
- agent-card defaultOutputModes: add text/plain — message/send always
  emits a kind:"text" part, so A2A v0.3.0 §5 wants it enumerated.

Verified: a2a/nlweb-ask/agent-mode-view/edge-functions 245/245,
typecheck:api + lint:api-contract + biome clean, docs stats unchanged.
